requirements.txt
allure-pytest==2.15.2
selenium==4.38.0
